requirements.txt
numpy
statistics
pandas
scipy
scikit-learn
category_encoders
django
djangorestframework
lime
django-pandas